+++ lcapy/expr.py
@@ -3859,6 +3859,37 @@ As a workaround use x.as_expr() %s y.as_expr()""" % op)
         result = foo(coeffs)
         return self.__class__(result, **self.assumptions)
 
+    def fit(self, x, y, method='trf', ranges=None, Ns=10, **kwargs):
+        """Find the parameters of an expression that best fits measured data using
+        ordinary least squares optimization techniques.
+
+        `x` is an ndarray of values for the dependent variable,
+        `y` is an ndarray of values for the independent variable,
+        `method` is the optimization method ('brute', 'trf' (default), or 'dogbox'),
+        `ranges` is a dictionary of the search ranges keyed by the parameter name,
+        `Ns` is the number of steps per range for the 'brute' optimizer.
+
+        The returned result is a `FitterResult` object.   This has attributes:
+        `params` a dictionary of the estimated parameter values keyed by parameter name,
+        `rmse` the root mean squared error.
+
+        Here's an example of use:
+        `
+        >>> e = expr('a * exp(-t  / tau) * u(t)')
+        >>> tv = arange(100)
+        >>> vv = e.subs({'a': 1, 'tau': 10}).evaluate(tv)
+        >>> results = e.fit(tv, vv, ranges={'a': (0, 10), 'tau': (1, 20)})
+        >>> results.params
+        {'a': 1.000000000048109, 'tau': 9.999999998432187}
+        >>> results.rmse
+        3.489526384702217e-22
+        `
+        """
+
+        from .fitter import fit
+
+        return fit(self, x, y, method, ranges, **kwargs)
+
     def force_time(self):
 
         if state.force_time: